Change in GUI
The new UI presents a text box to enter the cricinfo url instead of the
match number so that the application can be used for all matches in cricinfo.
Hence the variable name is not "matchno" anymore but "scoreurl"
The title has been changed to "Cricinfo Score App".
A border has been added around the area score is displayed.
The button and the window form have been made a little bit longer.
The text of the button is also changed1