New install and remove scripts.
This commit removes the ubuntuInstallation.sh script and replaces it
with install.sh and also modifies the remove.sh
Now the cricinfo app files are present in /opt/cricinfo folder and the
links places in /usr/bin are cricinfoscore and cricinfoscore-uninstall.
gui.py does not require any executable permissions as its used just for
import.
The installation of dependencies is removed from the install script to
make the installation portable. The required packages have been
mentioned in the README